+++ robotic_factory/src/model/mod.rs
@@ -1,3 +1,2 @@
-mod order;
 mod robot_factory;
 mod stations;
